Fix incorrect instructions in pokerhands exercise
One example case in pokerhands exercise states that a poker hand
"2S 8S AS QS 3S" wins over poker hand "2H 4S 4C 2D 4H". This is, however,
a mistake as the first-mentioned hand is a full house and the latter
is a flush. A full house wins always over a flush as per instructions.